+++ core/src/test/java/org/infinispan/util/concurrent/BoundedConcurrentHashMapTest.java
@@ -133,41 +133,6 @@ public class BoundedConcurrentHashMapTest extends EquivalentHashMapTest {
 //      }
 //   }
 
-   private void testRemovePerformance(Eviction eviction)
-   {
-      final int COUNT = 200000;
-      Map<Integer, Integer> bchm = new BoundedConcurrentHashMap<Integer, Integer>(
-            COUNT, 1, eviction, AnyEquivalence.INT, AnyEquivalence.INT);
-
-      // fill the cache
-      for (int i = 0; i < COUNT; i++)
-         bchm.put(i, i);
-
-      // force a single cache hit (so that accessQueue has a head item)
-      bchm.get(0);
-
-      // remove items
-      long start = System.currentTimeMillis();
-      for (int i = 1; i < COUNT; i++)
-      {
-         bchm.get(i);
-         bchm.remove(i);
-
-         // original version needs ~5 min for 200k entries (2h for 1M)
-         // fixed version needs < 200ms for 200k (500ms for 1M)
-         if (System.currentTimeMillis() - start > 5000)
-            fail(eviction.name() + ": removing " + COUNT + " entries takes more than 5 seconds!");
-      }
-   }
-
-   public void testLRURemovePerformance() {
-      testRemovePerformance(Eviction.LRU);
-   }
-
-   public void testLIRSRemovePerformance() {
-      testRemovePerformance(Eviction.LIRS);
-   }
-
    public void testLRUCacheHits() throws InterruptedException
    {
       final int COUNT_PER_THREAD = 100000;
